+++ ViewDeckExample/LeftViewController.m
@@ -86,7 +86,7 @@
 }
 
 - (NSString *)tableView:(UITableView *)tableView titleForHeaderInSection:(NSInteger)section {
-    return section > 0 ? [NSString stringWithFormat:@"%d", section-1] : nil;
+    return section > 0 ? [NSString stringWithFormat:@"%ld", (long)(section-1)] : nil;
 }
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
@@ -104,7 +104,7 @@
     }
     else {
         cell.accessoryType = UITableViewCellAccessoryDisclosureIndicator;
-        cell.textLabel.text = [NSString stringWithFormat:@"%d:%d", indexPath.section-1, indexPath.row];
+        cell.textLabel.text = [NSString stringWithFormat:@"%ld:%ld", (long)(indexPath.section-1), (long)indexPath.row];
     }
     return cell;
 }
